Move Intel-specific vcpu code to hve
Downstream projects like the extended APIs need a stable interface. To
facilitate this, the existing interface will be changed to pass a single
vcpu pointer to client code. Any state or functionality needed by
extensions will be reachable from the vcpu pointer.

This design requires forward declarations of the vcpu class, resulting
in circular references. While this works on Linux, Windows does not
allow inter-dll circular references.  The workaround is to place the
API's vcpu code into hve.

Move the Intel-specific vcpu code to hve to remain consistent
with the extended APIs.
