+++ bfvmm/src/hve/CMakeLists.txt
@@ -21,17 +21,17 @@ if(${BUILD_TARGET_ARCH} STREQUAL "x86_64")
         arch/intel_x64/check/check_vmcs_control_fields.cpp
         arch/intel_x64/check/check_vmcs_guest_fields.cpp
         arch/intel_x64/check/check_vmcs_host_fields.cpp
-        arch/intel_x64/vmx/vmx.cpp
-        arch/intel_x64/vmcs/vmcs.cpp
         arch/intel_x64/exit_handler/exit_handler.cpp
+        arch/intel_x64/vmcs/vmcs.cpp
+        arch/intel_x64/vmx/vmx.cpp
     )
 
     if(NOT WIN32 AND NOT ENABLE_MOCKING)
         list(APPEND SOURCES
+            arch/intel_x64/exit_handler/exit_handler_entry.asm
             arch/intel_x64/vmcs/vmcs_launch.asm
             arch/intel_x64/vmcs/vmcs_promote.asm
             arch/intel_x64/vmcs/vmcs_resume.asm
-            arch/intel_x64/exit_handler/exit_handler_entry.asm
         )
     endif()
 elseif(${BUILD_TARGET_ARCH} STREQUAL "aarch64")
